refactor: 💡 rename chat page body and refactor chat thread
section from chat page body

✅ Closes: #8
