+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "FAITH",
-  "version": "0.1.0+6",
+  "version": "0.1.0+7",
   "private": false,
   "scripts": {
     "dev": "next dev",
