+++ src/components/chat_message/chat_message.tsx
@@ -1,19 +1,16 @@
-import { MessageRole } from '@/interfaces/chat';
+import { MessageRole, Sender } from '@/interfaces/chat';
 import Image from 'next/image';
 import React from 'react';
 
 interface ChatMessageProps {
+  sender: Sender;
   role: MessageRole;
   content: string;
 }
 
-interface RoleMessageProps {
-  content: string;
-}
-
-const AnonymousUserMessage = ({ content }: RoleMessageProps) => {
-  return (
-    <div className="flex gap-5 self-start font-barlow">
+const ChatMessage = ({ sender, role, content }: ChatMessageProps) => {
+  const displayedAvatar =
+    role === MessageRole.VISITOR ? (
       <div className="relative flex shrink-0 items-start justify-center">
         <Image
           src={'/elements/anonymous_avatar.svg'}
@@ -22,43 +19,36 @@ const AnonymousUserMessage = ({ content }: RoleMessageProps) => {
           alt="isunfa logo with ring border"
         />
       </div>
-      <div className="flex flex-col">
-        {/* TODO: i18n (20240626 - Shirley) */}
-        <div className="text-xl font-bold leading-8">You</div>
-        <div className="mt-2 whitespace-pre-wrap text-base leading-6 tracking-normal">
-          {content}
-        </div>
+    ) : role === MessageRole.BOT ? (
+      <div className="relative flex h-60px w-60px shrink-0 items-center justify-center">
+        <Image
+          src={'/logo/isunfa_pure_logo.svg'}
+          width={30}
+          height={30}
+          alt="isunfa logo"
+          className="z-10 h-30px w-30px"
+        />
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="60"
+          height="60"
+          fill="none"
+          viewBox="0 0 60 60"
+          className="absolute"
+        >
+          <circle cx="30" cy="30" r="29" stroke="#CDD1D9" strokeWidth="2"></circle>
+        </svg>
       </div>
-    </div>
-  );
-};
+    ) : (
+      <p>user</p>
+    );
 
-const BotMessage = ({ content }: RoleMessageProps) => {
-  return (
+  const displayedMessage = (
     <div>
       <div className="mt-0 flex gap-5 self-start whitespace-nowrap font-barlow">
-        <div className="relative flex h-60px w-60px shrink-0 items-center justify-center">
-          <Image
-            src={'/logo/isunfa_pure_logo.svg'}
-            width={30}
-            height={30}
-            alt="isunfa logo"
-            className="z-10 h-30px w-30px"
-          />
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="60"
-            height="60"
-            fill="none"
-            viewBox="0 0 60 60"
-            className="absolute"
-          >
-            <circle cx="30" cy="30" r="29" stroke="#CDD1D9" strokeWidth="2"></circle>
-          </svg>
-        </div>
-
+        {displayedAvatar}
         <div className="flex w-1200px flex-col overflow-x-auto">
-          <div className="text-xl font-bold leading-8">Faith</div>
+          <div className="text-xl font-bold leading-8">{sender}</div>
           <div className="mt-2 whitespace-pre-wrap text-base leading-6 tracking-normal">
             {content}
           </div>
@@ -66,17 +56,6 @@ const BotMessage = ({ content }: RoleMessageProps) => {
       </div>
     </div>
   );
-};
-
-const ChatMessage = ({ role, content }: ChatMessageProps) => {
-  const displayedMessage =
-    role === MessageRole.ANONYMOUS_USER ? (
-      <AnonymousUserMessage content={content} />
-    ) : role === MessageRole.BOT ? (
-      <BotMessage content={content} />
-    ) : (
-      <div></div>
-    );
 
   return displayedMessage;
 };
