+++ src/interfaces/chat.ts
@@ -13,10 +13,16 @@ export interface IChatBrief {
   createdAt: number;
 }
 
+export enum Sender {
+  VISITOR = 'You',
+  USER = '',
+  FAITH = 'Faith',
+}
+
 export enum MessageRole {
   USER = 'user',
   BOT = 'bot',
-  ANONYMOUS_USER = 'anonymous user',
+  VISITOR = 'visitor',
 }
 
 export interface IMessage {
@@ -52,7 +58,7 @@ export const dummyMessage: IMessage = {
 export const dummyMessages: IMessage[] = [
   {
     id: 'masvdsf',
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '你能解釋一下什麼是遞迴函數嗎？',
     createdAt: Date.now(),
   },
@@ -64,7 +70,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '可以舉個例子嗎？',
     createdAt: Date.now(),
   },
@@ -83,7 +89,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '這個函數是如何工作的？',
     createdAt: Date.now(),
   },
@@ -96,7 +102,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '遞迴函數有什麼優缺點？',
     createdAt: Date.now(),
   },
@@ -109,7 +115,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '還有其他常見的遞迴函數例子嗎？',
     createdAt: Date.now(),
   },
@@ -122,7 +128,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '能詳細解釋一下斐波那契數列的遞迴實現嗎？',
     createdAt: Date.now(),
   },
@@ -135,7 +141,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '遞迴和迭代有什麼區別？',
     createdAt: Date.now(),
   },
@@ -148,7 +154,7 @@ export const dummyMessages: IMessage[] = [
   },
   {
     id: uuidv4(),
-    role: MessageRole.ANONYMOUS_USER,
+    role: MessageRole.VISITOR,
     content: '如何避免遞迴導致的堆棧溢出？',
     createdAt: Date.now(),
   },
