+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import ChatSidebar from '@/components/chat_sidebar/chat_sidebar';
 import NavBar from '@/components/nav_bar/nav_bar';
-import VisitorChatPageBody from '@/components/visitor_chat_page_body/visitor_chat_page_body';
+import ChatPageBody from '@/components/chat_page_body/chat_page_body';
 import Head from 'next/head';
 import { useState } from 'react';
 
@@ -41,7 +41,7 @@ export default function Home() {
             className={`flex h-screen flex-col justify-end transition-all duration-300 ease-in-out ${isSidebarExpanded ? 'ml-240px' : ''}`}
           >
             {' '}
-            <VisitorChatPageBody />
+            <ChatPageBody />
           </div>
         </div>
       </div>
